Move shared grant helpers out of team_grants.go
LocationGrant, grantScopeToAPI, and scopeForGrant are used by all three
principal grant client files. Extract them to a neutral grants.go so the
file naming no longer suggests ownership by teams.
